Changed TASK_GENERATOR_SERVER_NAME relative -> global
This was a bug which broke the rosnav agent testing. If the rosnav node were to access the task generator parameter server, the service name would be resolved to /[model name]/task_generator_node/set_parameters since rosnav node is launched in the namespace [model name]. To prevent this, I made the constant a global name. This way, the namespace is not added when resolving the name.
I believe this has no effect on other testing and training since the namespace is always task_generator_server.